🚚 Rename plugin to bridgetown_notion
Also deleted unecessary scaffolded files and folders.
