--- components/bridgetown_notion/layout_help.liquid
@@ -1,1 +1,0 @@
-<p>Just use <code>layout: bridgetown_notion/layout</code> in your page's front matter.</p>
