--- content/bridgetown_notion/example_page.md
@@ -1,8 +1,0 @@
----
-layout: default
-title: Example Page
----
-
-If all goes well, this page will be accessible as [`{{ page.url }}`]({{ page.url }}), and you will see a photo of a train:
-
-![Train on Rails]({{ "/bridgetown_notion/train-on-rails.jpeg" | relative_url }})
