--- frontend/javascript/index.js
@@ -1,10 +1,0 @@
-/* Use in your app by simply adding to your app's index.js:
-
-import BridgetownNotion from "bridgetown-notion"
-
-BridgetownNotion()
-*/
-
-export default function() {
-  console.log("Success! The sample plugin has been loaded.")
-}
