--- frontend/styles/index.scss
@@ -1,8 +1,0 @@
-/* Use in your app by simply adding to your app's index.scss:
-
-@import "~bridgetown-sample-plugin/frontend/styles";
-*/
-
-body {
-  background: #ffc;
-}
