--- layouts/bridgetown_notion/layout.html
@@ -1,9 +1,0 @@
----
-layout: default
----
-
-<p>You can include this layout in your site! 🥳</p>
-
-{% render "bridgetown_notion/layout_help" %}
-
-{{ content }}
